requirements.txt
python-dotenv==1.1.0
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.31.0
schedule==1.2.1